Init GameEvents.
This is broker because the current working approach of TrackerEvents
considers that one EventId tags one actual EventType.
So we can do a match on a tag, and then it follows the Event itself.

In the case of GameEvents, there are multiple tags per EventType.
We need to figure out what to do here.

Signed-off-by: Seb Ospina <kraige@gmail.com>