ConstDecl skipping for code generation
There are references to Constants in struct fields, such references
explain in the current protocol what values are hardcoded. These values
then are not encoded in the events and so there shouldn't be parsing for
them.
It has also been observed that when such ConstDecl (i.e. maxColors) are
later referenced, they are accompanied by a `.value` in the json, so
while knowing such a max value points to a "max color" value, it also
has its own `.max.evalue` field would contain what is the maximum value
that we are pointing at.

This max value is later used to calculate the number of bits needed to
read to get the full value. This can be inclusive, when such a case is
found, 1 is added to the number of bits needed from the sqrt()
calculation.

Signed-off-by: Seb Ospina <kraige@gmail.com>